feat:提交返点比率模板相关接口

+++ src/main/java/com/dx/facade/member/param/AddRebateRateTemplateMappingParamDTO.java
@@ -35,10 +35,10 @@ public class AddRebateRateTemplateMappingParamDTO implements Serializable {
     private Long userId;
     
     @ApiModelProperty("用户类型 0：会员  1：代理")
-    private Long userType;
+    private Integer userType;
     
     @ApiModelProperty("总代Id")
-    private Long topProxyId;
+    private Integer topProxyId;
     
     @ApiModelProperty("创建人")
     private String createdBy;
